requirements-tests.txt
-r requirements.txt
tox>=1.8
django-discover-runner
djangocms-helper>=0.7
